pyproject.toml
[project]
name = "sat-sn"
version = "0.1.0"
description = ""
authors = [
    {name = "pedroKomcorp",email = "pedro.neto@komcorp.com.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdfplumber (>=0.11.7,<0.12.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "camelot-py[base] (>=1.0.0,<2.0.0)",
    "fitz (>=0.0.1.dev2,<0.0.2)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "itables (>=2.4.4,<3.0.0)",
    "streamlit (>=1.48.0,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "python-calamine (>=0.4.0,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

